agent/brain/python/pyproject.toml
[project]
name = "python_haxball_ai_brain"
version = "0.1.0"
description = "A stupid python haxball_ai brain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mypy>=1.17.0",
    "nats-py>=2.10.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
